# Add a for each method for visiting query results due to its superior performance compared to the iterator variant.