I am somewhat surprised at how much better the code gen is:
```
test query_single_archetype                    ... bench:      47,422 ns/iter (+/- 469)
test query_for_each_single_archetype           ... bench:      16,677 ns/iter (+/- 355)

test query_many_archetypes                     ... bench:      47,881 ns/iter (+/- 639)
test query_for_each_many_archetypes            ... bench:      13,714 ns/iter (+/- 4,037)

test query_very_many_small_archetypes          ... bench:       1,213 ns/iter (+/- 38)
test query_for_each_very_many_small_archetypes ... bench:         863 ns/iter (+/- 7)
```